Remove Test Snapshots from .gitignore
This was a bit of an oopsy; a major component of utilizing snapshots with Jest is to ensure you are committing the snapshot (or snapshot updates) with the associated code changes. From [Jest’s official docs on snapshot testing](https://jestjs.io/docs/en/snapshot-testing):

> The snapshot artifact should be committed alongside code changes, and reviewed as part of your code review process.

Similarly; snapshots must be updated as we update the associated views for each area of the application.